+++ projects/media-viewer/src/lib/media-viewer/toolbar/main-toolbar/main-toolbar.component.html
@@ -11,8 +11,8 @@
                          [stepZoomEvent]="actionEvents?.stepZoom"
                          [rotateEvent]="actionEvents?.rotate"
                          [zoomValue]="(actionEvents?.zoomValue | async) || { value: 1 }"
-                         [showZoomBtns]="toolbarToggles.showZoomBtns | async"
-                         [showRotateBtns]="toolbarToggles.showRotateBtns | async"></mv-tb-middle-pane>
+                         [showZoomBtns]="toolbarToggles?.showZoomBtns | async"
+                         [showRotateBtns]="toolbarToggles?.showRotateBtns | async"></mv-tb-middle-pane>
     </div>
     <div id="loadingBar">
       <div class="progress">
